Remove python 3 compat imports. Rename logging.py to avoid conflicts with python logging module.
