+++ src/appleseed.python/CMakeLists.txt
@@ -83,7 +83,7 @@ set (sources
     dict2dict.h
     gillocks.h
     gillocks.cpp
-    logging.py
+    logtarget.py
     metadata.h
     module.cpp
     studio.py
@@ -163,7 +163,7 @@ target_link_libraries (appleseed.python
 
 add_copy_target_to_sandbox_py_module_command (appleseed.python)
 add_copy_py_file_to_sandbox_py_module_command (${PROJECT_SOURCE_DIR}/src/appleseed.python/__init__.py)
-add_copy_py_file_to_sandbox_py_module_command (${PROJECT_SOURCE_DIR}/src/appleseed.python/logging.py)
+add_copy_py_file_to_sandbox_py_module_command (${PROJECT_SOURCE_DIR}/src/appleseed.python/logtarget.py)
 add_copy_py_file_to_sandbox_py_module_command (${PROJECT_SOURCE_DIR}/src/appleseed.python/studio.py)
 add_copy_dir_to_sandbox_py_module_command (${PROJECT_SOURCE_DIR}/src/appleseed.python/test)
 
@@ -174,7 +174,7 @@ add_copy_dir_to_sandbox_py_module_command (${PROJECT_SOURCE_DIR}/src/appleseed.p
 
 set (py_module_dst "lib/python${PYTHON_MAJOR_VERSION}.${PYTHON_MINOR_VERSION}/appleseed")
 install (FILES __init__.py DESTINATION ${py_module_dst})
-install (FILES logging.py DESTINATION ${py_module_dst})
+install (FILES logtarget.py DESTINATION ${py_module_dst})
 install (FILES studio.py DESTINATION ${py_module_dst})
 install (DIRECTORY test DESTINATION ${py_module_dst})
 install (TARGETS appleseed.python DESTINATION ${py_module_dst})
