+++ src/appleseed.python/logtarget.py
